PHPStan: Level 2
OrderShippingExtension.php: in `createShippingPackage` method, change casting of $value from int to float to avoid the loss of precision upon conversion
ZoneSelectField.php: use methods of Map class to populate select field
DistanceShippingMethod.php, TableShippingMethod.php, Zone.php, ZonedShippingMethod.php:  call method `removeByName` on DataList not Field List
ShippingFrameworkModifier.php: ensure that order and shipping method exists before calling methods
ShippingEstimatorTest.php: use SilverStripe method `::create` rather than using the new keyword
TableShippingMethodTest.php: restyle array

+++ src/Extension/OrderShippingExtension.php
@@ -39,7 +39,7 @@ class OrderShippingExtension extends Extension
     /**
      * Create package, with total weight, dimensions, value, etc.
      */
-    public function createShippingPackage(int $value = 0): ShippingPackage
+    public function createShippingPackage(float $value = 0): ShippingPackage
     {
         $items = $this->owner->Items();
